Enable -Wall -Wextra -Wno-unused-parameter on GCC/Clang.
This is good practice and helps to catch bugs. Several changes
were made to accomodate the newly enabled warnings:
  * -Wunused-function:
    * in exposed/, static functions that were supposed to be inlined
      were explicitly marked as inline;
    * some actually unused functions were removed;
  * -Wsign-compare: explicit conversions were added, and in
    the future we should find a nicer way than aux* fields;
  * -Wmissing-field-initializers: added initializers;
  * -Wreorder: reordered properly;
  * -Wunused-but-set-variable: remove variable.

-Wunused-parameter was turned off as enabling it would result in
massive amount of churn in UI code. Despite that, we should enable
it at some point as it has a fairly high SNR otherwise.

+++ src/exportvector.cpp
@@ -182,7 +182,7 @@ public:
             bool used = false;
             for(DxfFileWriter::BezierPath &path : writer->paths) {
                 for(SBezier *sb : path.beziers) {
-                    if(sb->auxA != s->h.v) continue;
+                    if((uint32_t)sb->auxA != s->h.v) continue;
                     used = true;
                     break;
                 }
